// src/lib/types/generated/ExamineeImportColumn.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ExamineeImportColumn =
	| 'subjectName'
	| 'rowIdentifier'
	| 'examineeNif'
	| 'examineeName'
	| 'examineeSurenames'
	| 'examineeOrigin'
	| 'examineeCourt'
	| 'examineeAcademicCentre';