// src/lib/types/generated/ExamineeImportError.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExamineeForImport } from './ExamineeForImport';
import type { ExamineeImportInvalidValueError } from './ExamineeImportInvalidValueError';
import type { ExamineeImportMissingValueError } from './ExamineeImportMissingValueError';
import type { ExamineeImportMissmatchError } from './ExamineeImportMissmatchError';

export type ExamineeImportError =
	| 'lock'
	| 'noValuesLoaded'
	| 'noSheet'
	| { missingValue: { row: number; missing: ExamineeImportMissingValueError } }
	| {
			missmatchValue: {
				row: number;
				missmatch: ExamineeImportMissmatchError;
				establishedValue: string;
				foundValue: string;
			};
	  }
	| { invalidValue: { row: number; reason: ExamineeImportInvalidValueError; invalidValue: string } }
	| { missingExamineeValue: { examinee: ExamineeForImport } };